+++ src/observerPattern/observables/ObservableAddress.java
@@ -3,7 +3,7 @@ package observerPattern.observables;
 import domainLogic.util.Address;
 
 public class ObservableAddress extends Address {
-    public ObservableAddress(Integer addressCount) {
+    public ObservableAddress(Long addressCount) {
         super(addressCount);
     }
 }
